Add fokkonauts Souly taser (closed #326)

+++ src/game/server/player_ddpp.h
@@ -318,7 +318,7 @@ public:
 	int m_ShopBotMesssagesRecieved;
 
 	void JailPlayer(int seconds);
-	int TaserFreezeTime();
+	float TaserFreezeTime();
 	int m_TaserPrice;
 	bool m_TaserOn;
 	bool m_IsJailed;
